Added more details to the preview:
- Text items are added for each server and for each IP address.
- Each server item shows how many IP addresses it contains.
- Each address shows the result of the NTP query (correction and latency).

Due to limitations of the Wii U Plugin System, the extra items are not shown until the
user either presses DOWN then UP (to try to scroll the category) or exit then re-enters
the preview category. Also, WUPS does not allow items to be removed; when a server returns
different addresses (e.g. pool.ntp.org) the old addresses will still be displayed.